% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AA_atlas}
\alias{AA_atlas}
\title{Exon boundaries for all exons in protein coding genes in the IPD-IMGT/HLA Database release v 3.39.0}
\format{
A list containing exon boundaries in a dataframe format for each locus.
}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
AA_atlas
}
\description{
A list object containing exon boundaries for all exons in protein coding genes in the IPD-IMGT/HLA Database release v 3.39.0. Exon boundaries were determined from the nucleotide alignment files, which were downloaded from the ANHIG/IMGTHLA Github respository.
}
\note{
For internal use only.
}
\keyword{datasets}
