% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAlleleSyntax.R
\name{checkAlleleSyntax}
\alias{checkAlleleSyntax}
\title{Check allele syntax}
\usage{
checkAlleleSyntax(allele, filename)
}
\arguments{
\item{allele}{An allele name written in the IPD-IMGT/HLA Database format.}

\item{filename}{The full file path of the user specified dataset if the user wishes to use their own file, or the pre-bundled Solberg dataset. User provided datasets must be a .dat, .txt, or.csv file, and must conform to the structure and format of the Solberg dataset.}
}
\value{
TRUE if allele syntax is correct. Otherwise,a vector containing FALSE and an error message is returned.
}
\description{
Checks if allele syntax is valid.
}
\note{
For internal SSHAARP use only.
}
\examples{
#Example where allele has correct number of fields where user specified dataset is the Solberg dataset
checkAlleleSyntax("A*01:01:01", filename=SSHAARP::solberg_dataset)

#Example where allele has incorrect number of fields where user specified dataset is the Solberg dataset
checkAlleleSyntax("A*01:01:01:01:01", filename=SSHAARP::solberg_dataset)

#Example where allele has incorrect syntax where user specified dataset is named "myfilename.dat"
\dontrun{checkAlleleSyntax("A*01:01:0:::", filename="myfilename.dat")}
}
