% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPosition.R
\name{checkPosition}
\alias{checkPosition}
\title{Checks if amino acid positions in motif exist}
\usage{
checkPosition(motif, filename)
}
\arguments{
\item{motif}{An amino acid motif in the following format: Locus*##$~##$~##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Motifs can include any number of amino acids. This function ONLY checks if the entered amino acid positions exist in IMGTprotalignments.}

\item{filename}{The full file path of the user specified dataset if the user wishes to use their own file, or the pre-bundled Solberg dataset. User provided datasets must be a .dat, .txt, or.csv file, and must conform to the structure and format of the Solberg dataset.}
}
\value{
TRUE if all of the amino acid positions in a motif exist. Otherwise, a vector with FALSE and an error message is returned.
}
\description{
Checks if amino acid positions in the entered motif exist in IMGTprotalignments.
}
\note{
For internal SSHAARP use only.
}
\examples{
#Example with existent amino acid positions
checkPosition("DRB1*26F~28E", filename=SSHAARP::solberg_dataset)

#Example with nonexistent amino acid positions
checkPosition("DRB1*199999F", filename=SSHAARP::solberg_dataset)
}
