% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_haplotype_dataset}
\alias{mock_haplotype_dataset}
\title{Mock Haplotype Dataset}
\format{
An object of class \code{data.frame} with 24989 rows and 7 columns.
}
\usage{
mock_haplotype_dataset
}
\description{
A dataframe containing mock haplotype data modeled after the Allele Frequency Network Database (AFND) haplotype data structure.
}
\keyword{datasets}
