% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solberg_dataset}
\alias{solberg_dataset}
\title{Solberg Dataset}
\format{
A dataframe with 20163 rows and 13 columns.
}
\source{
results.zip file from http://pypop.org/popdata/
}
\usage{
solberg_dataset
}
\description{
A dataframe of the original Solberg dataset, which is a global dataset of 497 population samples from 185 published datasets, representing 66,800 individuals.
For more information on the Solberg dataset, please see the vignette.
}
\references{
Solberg et.al "Balancing selection and heterogeneity across the classical human leukocyte antigen loci: A meta-analytic review of 497 population studies". Human Immunology (2008) 69, 443–464
}
\keyword{datasets}
