% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateVariantType.R
\name{validateVariantType}
\alias{validateVariantType}
\title{validateVariantType}
\usage{
validateVariantType(vT)
}
\arguments{
\item{vT}{Specifies whether the variant is an allele, motif, or haplotype.}
}
\value{
TRUE if the entered variantType is accepted ('motif', 'allele', or 'haplotype'). FALSE for any other provided entry.
}
\description{
Validates the variantType parameter in PALM()
}
\note{
For internal SSHAARP use only
}
