% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyAlleleANHIGHaplo.R
\name{verifyAlleleANHIGHaplo}
\alias{verifyAlleleANHIGHaplo}
\title{Verifies the alleles in entered haplotype are present in IMGTprotalignments}
\usage{
verifyAlleleANHIGHaplo(haplotype, filename)
}
\arguments{
\item{haplotype}{A haplotype where allele names are written in the IPD-IMGT/HLA Database format, and have 1-4 fields. Alleles in haplotypes may be delimited by "-" or "~".}

\item{filename}{The full file path of the user specified dataset if the user wishes to use their own file, or pre-bundled mock haplotype dataset. User provided datasets must be a .dat, .txt, or.csv file, and must conform to the structure and format of the mock haplotype dataset bundled with the package.}
}
\value{
TRUE if all alleles in a haplotype are present in the IMGTprotalignment object. Otherwise, a vector containing FALSE and an error message is returned.
}
\description{
Verifies the alleles in entered haplotype are present in IMGTprotalignments.
}
\note{
For internal SSHAARP use only.
}
\examples{
#Example of a haplotype where all alleles are present in the IMGTprotalignment object
\dontrun{verifyAlleleANHIGHaplo("DRB1*01:01~A*01:01", filename=mock_haplotype_dataset)}

#Example of a haplotype where an allele is not present in the IMGTprotalignment object
\dontrun{verifyAlleleANHIGHaplo("DRB1*01:01~A*01:9999", filename=mock_haplotype_dataset)}
}
