% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantInfo.R
\name{getVariantInfo}
\alias{getVariantInfo}
\title{Locus and allele or motif extraction for motifs, allele, or haplotype mapping}
\usage{
getVariantInfo(variant)
}
\arguments{
\item{variant}{An amino acid motif or allele. Amino acid motifs must be in the following format: Locus*##$~##$~##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Alleles must have 1-4 fields.}
}
\value{
A list object with loci and allele or motif information.
}
\description{
Extracts locus and allele or motif information from variant.
}
\note{
For internal SSHAARP use only.
}
