% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMGTprotalignments}
\alias{IMGTprotalignments}
\title{Protein alignments for all protein coding genes in the IPD-IMGT/HLA Database release v 3.39.0.}
\format{
A list containing protein alignments in a dataframe format for each locus.
}
\source{
https://github.com/ANHIG/IMGTHLA/tree/Latest/alignments
}
\usage{
IMGTprotalignments
}
\description{
A list object containing protein alignments for all protein coding genes in the IPD-IMGT/HLA Database release. Alignments were downloaded from the ANHIG/IMGTHLA Github respository.
}
\keyword{datasets}
