% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isNamePresent.R
\name{isNamePresent}
\alias{isNamePresent}
\title{Checks if name portion of entered variant is present}
\usage{
isNamePresent(variant, variantType)
}
\arguments{
\item{variant}{An amino acid motif or allele. Amino acid motifs must be in the following format: Locus*##$~##$~##$, where ## identifies a peptide position, and $ identifies an amino acid residue. Alleles must have 1-4 fields.}

\item{variantType}{Identifies whether the variant is an allele or motif.}
}
\value{
TRUE if name is present. Otherwise, a vector with FALSE and an error message is returned.
}
\description{
Checks if the name portion of the entered variant is present. Names consist of information following the locus and asterisk of the entered variant.
}
\note{
For internal SSHAARP use only.
}
